Finer interploation for sun-shape sampling

+++ ClassedSunRay/recthelio_tracing.h
@@ -3,7 +3,7 @@
 #include "steps_for_raytracing.h"
 #include "ray_tracing.h"
 
-float recthelio_ray_tracing_init(const RectangleHelio &recthelio,			//	which heliostat will be traced
+int recthelio_ray_tracing_init(const RectangleHelio &recthelio,			//	which heliostat will be traced
 								const Grid &grid,							//	the grid heliostat belongs to
 								const vector<Heliostat *> heliostats,		//	all heliostats
 								const int &num_group,						//	number of sun-ray
@@ -13,7 +13,7 @@ float recthelio_ray_tracing_init(const RectangleHelio &recthelio,			//	which hel
 								float3 *&d_helio_vertexs,
 								int *&d_microhelio_groups);
 
-float recthelio_ray_tracing(const SunRay &sunray,
+int recthelio_ray_tracing(const SunRay &sunray,
 							Receiver &receiver,
 							const RectangleHelio &recthelio,		//	which heliostat will be traced
 							Grid &grid,								//	the grid heliostat belongs to
